Update Community.js
Used working Discord invite link to correct broken Discord invite link

Please ensure these are intended to be the same link. Else you will need to generate a new invite link from Discord to account for this. But I believe they are intended to be the same and were overlooked.